requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alabaster==0.7.16
asgiref==3.7.2
astroid==3.0.3
async-timeout==4.0.2
attrs==23.1.0
autopep8==1.6.0
babel==2.16.0
black==24.3.0
bleach==4.1.0
certifi==2021.10.8
cfgv==3.4.0
charset-normalizer==2.0.7
click==8.0.3
codespell==2.1.0
colorama==0.4.6
coverage==7.6.1
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.8
docutils==0.20.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
filelock==3.15.4
flake8==6.1.0
Flask==2.0.2
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.41
httpretty==1.1.4
identify==2.6.0
idna==3.3
imagesize==1.4.1
importlib_metadata==8.2.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.4.2
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
mypy==1.9.0
mypy-extensions==1.0.0
nh3==0.2.18
nodeenv==1.9.1
numpy==2.1.3
openai==0.19.0
openpyxl==3.0.9
-e git+https://github.com/sundeep110/opentelemetry-python.git@b2b133dcdadd05e3f90603fc27ecc4783d43d5f2#egg=opentelemetry_api&subdirectory=opentelemetry-api
-e git+https://github.com/sundeep110/opentelemetry-python.git@b2b133dcdadd05e3f90603fc27ecc4783d43d5f2#egg=opentelemetry_sdk&subdirectory=opentelemetry-sdk
-e git+https://github.com/sundeep110/opentelemetry-python.git@b2b133dcdadd05e3f90603fc27ecc4783d43d5f2#egg=opentelemetry_semantic_conventions&subdirectory=opentelemetry-semantic-conventions
packaging==24.1
pandas==1.3.4
pandas-stubs==1.2.0.35
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.0
psutil==5.9.6
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.18.0
pylint==3.0.2
pytest==7.4.4
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0.2
readme-renderer==42.0
requests==2.32.3
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
Sphinx==7.1.2
sphinx-autodoc-typehints==1.25.2
sphinx-rtd-theme==2.0.0rc4
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.2
tqdm==4.62.3
typing_extensions==4.12.2
urllib3==1.26.7
virtualenv==20.26.3
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.16.0
yarl==1.9.2
zipp==3.20.0
